export from index.js and factor out into is-leveldown.js
